django/001-django-rest/requirements.txt
django==5.0.1
djangorestframework==3.14.0
mysqlclient==2.2.3